apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
httpx==0.27.2
PyJWT==2.9.0
cryptography>=42
reportlab>=4
feedparser>=6
pypdf>=4
